memory safety improvements

+++ tools/source_maker/source_templates/references.cpp
@@ -61,8 +61,7 @@ void wxPHPObjectReferences::RemoveReferences()
         php_printf("Removing References\n");
         #endif
 
-        for(unsigned int i=0; i<m_references.size(); i++)
-        {
+        for (size_t i = 0; i < m_references.size(); i++) {
             if (Z_REFCOUNTED_P((m_references[i])) && Z_TYPE_P(m_references[i]) > IS_UNDEF && Z_TYPE_P(m_references[i]) <= _IS_ERROR) {
                 #ifdef USE_WXPHP_DEBUG
                 php_printf("Removing reference: %i\n", i);
